feat: tree sitter context (#10)
* feat: use tree-sitter to parse comments and their context

* refactor: error early when determing file type

It's better to return an error right away instead of carrying the
`Unknown` variant around the code.

* feat: add citation context to python bindings

* test: clean-up tests after merge

The tests from main were asserting individual fields, but we can rely on
`Annotation` equality to check everything together.

* refactor: rename CitationContext to SyntaxContext

* feat: determine whether annotation comment is inline

* feat: stop annotation context parsing on closest class too

We weren't breaking on the class, only in function definitions. We're
now stopping on both.